pyproject.toml
[tool.poetry]
name = "quicksummarizer"
version = "0.1.0"
description = ""
authors = ["Guillaume Demay <guillaumedemay@hotmail.fr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.95.0"
mangum = "^0.17.0"
uvicorn = "^0.21.1"
aws-sam-cli = "^1.78.0"


[tool.poetry.group.dev.dependencies]
openai = "^0.27.2"
python-dotenv = "^1.0.0"
httpx = "^0.23.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
